backend/requirements.txt
# --- Web & API ---
fastapi
uvicorn
python-dotenv
requests
pydantic

# --- AI & Frameworks ---
openai
langchain
langchain-community
langchain-openai
langchain-core
langgraph
mcp

# --- Tools & Scraping ---
duckduckgo-search
feedparser
beautifulsoup4  # (This creates the 'bs4' folder you saw)

# --- Data, Vision & RAG (The "Heavy" Stuff) ---
chromadb        # (This pulled in torch, transformers, posthog, etc.)
scikit-learn    # (This creates the 'sklearn' folder)
numpy
pillow          # (This creates the 'PIL' folder)

# --- Logistics ---
geopy